fix: 설문 엔티티와 DB 숫자 타입 정합성 수정

+++ src/main/java/com/planit/domain/PreferenceQuestion.java
@@ -25,7 +25,7 @@ public class PreferenceQuestion {
     private String questionText;
 
     @Column(name = "display_order", nullable = false)
-    private int displayOrder;
+    private short displayOrder;
 
     protected PreferenceQuestion() {
     }
@@ -34,7 +34,7 @@ public class PreferenceQuestion {
         return id;
     }
 
-    public int getDisplayOrder() {
+    public short getDisplayOrder() {
         return displayOrder;
     }
 }
